requirements.txt
alembic==1.12.0
black==23.10.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.1
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
flake8==6.1.0
Flask==2.3.3
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mccabe==0.7.0
mlflow==2.7.1
mypy-extensions==1.0.0
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
protobuf==4.24.4
pyarrow==13.0.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
tabulate==0.9.0
threadpoolctl==3.2.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
websocket-client==1.6.4
Werkzeug==3.0.1
zipp==3.17.0
